pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "opencv-python-headless>=4.13.0.90",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "plotly>=6.5.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.53.1",
    "tensorflow>=2.20.0",
    "tf-keras-vis>=0.8.7",
]
